refactor: CustomModal & FoodList 컴포넌트 모듈화 완료 (#89)
* feat: env 설정 추가 완료

* chore: dotenv-webpack 추가 완료

* feat: env 설정 추가 완료

* feat: Spinner 컴포넌트 구현 완료

* feat: 페이지 기본 설정 완료

* feat: route 상수 설정 완료

* feat: route 추가 완료

* feat: Router 설정 완료

* feat: Suspense & lazy 기본 설정 완료

* feat: GlobalNav 컴포넌트 기본 설정 완료

* feat: 파비콘 & 타이틀 기본 설정 완료

* feat: favicon 설정 추가 완료

* fix: 경로 수정 완료

* design: GlobalNav 컴포넌트 UI & 반응형 CSS 구현 완료

* feat: GlobalNav 컴포넌트 구현 완료

* feat: color 상수 설정 완료

* feat: screen 상수 설정 완료

* refactor: 객체 요소로 변경 완료

* feat: color & screen 추가 완료

* refactor: 경로 네이밍 리펙토링 완료

* feat: button 기본 설정 추가 완료

* feat: BASKET 경로 추가 완료

* feat: Basket 페이지 경로 추가 완료

* feat: Basket 컴포넌트 기본 설정 완료

* design: 버튼 UI 추가 완료

* feat: 장바구니 버튼 추가 완료

* design: Login 컴포넌트 기본 UI 설정 완료

* feat: Login 컴포넌트 와이어프레임 구현 완료

* feat: axios 기본 설정 완료

* feat: 카카오 Oauth 2.0 로그인 구현 로직 완료

* feat: 카카오 Oauth 2.0 로그인 구현 로직 완료

* feat: api 상수 설정 완료

* feat: api 추가 완료

* chore: 네트리파이 환경 변수 설정 완료

* feat: systemvars 설정 추가 완료

* fix: 환경 변수 설정 삭제

* feat: 폰트 추가 완료

* feat: 기본 폰트 변경 완료

* design: reset 폰트 삭제

* design: Login 컴포넌트 UI & CSS 구현 완료

* feat: IDINQUIRY & PWINQUIRY 프로퍼티 추가 완료

* feat: API_URI 상수 추가 완료

* feat: API_URI 프로퍼티 추가 완료

* feat: 회원가입 폼 입력 정제 완료

* feat: 카카오 로그인 로직 구현 완료

* docs: 배포 링크 & 주제 선정 이유 추가 완료

* feat: IdInquiry 컴포넌트 와이어프레임 구현 완료

* feat: PwInquiry 와이어프레임 구현 완료

* design: Form 기본 UI 설정 완료

* feat: Thumbnail 컴포넌트 구현 완료

* design: Thumbnail UI & CSS 구현 완료

* design: 회원가입 버튼 UI 추가 완료

* desgin: Home 컴포넌트 기본 UI 구현 완료

* feat: REGISTER 프로퍼티 추가 완료

* design: Home 컴포넌트 UI & CSS 구현 완료

* design: Thumbnail 컴포넌트 UI & 반응형 CSS 구현 완료

* feat: LocalNav 컴포넌트 기본 설정 완료

* refactor: 사용하지 않는 프로퍼티 제거 완료

* design: Register 와이어프레임 구현 완료

* design: LocalNav 컴포넌트 기본 UI 구현 완료

* design: LocalNav 와이어프레임 구현 완료

* design: 반응형 CSS 추가 완료

* feat: 네비게이션 기능 추가 완료

* feat: location 데이터 정제 완료

* feat: menu 상수 설정 완료

* feat: menu 추가 완료

* feat: data 속성 query 추가 완료

* refactor: Banner 컴포넌트로 분리 완료

* design: 배너 & 네비게이션 바 UI 제공 완료

* design: Banner 컴포넌트 UI & 반응형 CSS 구현 완료

* refactor: Banner 컴포넌트 분리 완료

* refactor: 유동 데이터 UI 랜더링 수정 완료

* fix: 중괄호 누락 수정 완료

* refactor: 컴포넌트 모듈화 완료

* refactor: 컴포넌트 import 변경 완료

* design: FoodList 컴포넌트 UI & CSS 구현 완료

* design: FoodTruck 컴포넌트 와이어프레임 구현 완료

* design: FoodTruck UI & 반응형 CSS 구현 완료

* feat: FoodTruck 컴포넌트 추가 완료

* feat: FoodList 컴포넌트 인터페이스 구현 완료

* fix: 컴포넌트 key 추가 완료

* feat: FoodList 중첩 라우터 기능 추가 완료

* feat: FoodTruck 네이게이션 기능 추가 완료

* feat: 경로 설정 완료

* feat: FoodTrucks 컴포넌트 구현 완료

* feat: menuQuery atom 생성 완료

* feat: atoms 저장소 구현 완료

* feat: menuQuery 전역 상태 관리 구현 완료

* feat: FOODLIST 프로퍼티 추가 완료

* refactor: 상수로 리펙토링 완료

* feat: #modal 엘리먼트 추가 완료

* design: Modal 컴포넌트 UI & CSS 구현 완료

* feat: createPortal 기능 구현 완료

* feat: FoodModal 컴포넌트 추가 완료

* refactor: Modal 컴포넌트 모듈화

* refactor: Modal 컴포넌트 모듈화

* feat: FoodModal 컴포넌트 구현중

* design: 컨텐츠 y축 auto 스크롤 인터페이스 구현 완료

* feat: OrderModal 컴포넌트 구현 완료

* feat: EmailModal 컴포넌트 구현 완료

* feat: props width, height 추가 완료

* feat: Modal 컴포넌트 추가 완료

* feat: Modal 컴포넌트 모듈화 완료

* feat: modal atom 생성 완료

* feat: useModal 구현 완료

* feat: useModal 추가 완료

* feat: 모달 테스트 인터페이스 제공 완료

* refactor: CSS 로직 분리 완료

* feat: EmailModal 컴포넌트 구현 완료

* feat: FoodModal 컴포넌트 구현 완료

* feat: OrderModal 컴포넌트 구현 완료

* refactor: CustomModal 컴포넌트 자식 컴포넌트 모듈화 완료

* refactor: 컴포넌트 이름 직관적이게 변경 완료

* refactor: 중복 스타일 제거 완료

* refactor: FoodList 컴포넌트 자식 컴포넌트 모듈화 완료

* fix: HTTP 메소드 수정 완료
