+++ frontend/src/components/CustomModal/index.js
@@ -2,10 +2,10 @@ import { FoodModal } from './FoodModal';
 import { OrderModal } from './OrderModal';
 import { EmailModal } from './EmailModal';
 
-const Modal = {
-  FoodModal,
-  OrderModal,
-  EmailModal,
+const CustomModal = {
+  Food: FoodModal,
+  Order: OrderModal,
+  Email: EmailModal,
 };
 
-export { Modal };
+export { CustomModal };
