+++ frontend/src/pages/FoodList/FoodDetail/styles.js
@@ -1,17 +1,5 @@
 import styled from 'styled-components';
 
-const FoodListContainer = styled.div`
-  max-width: 1020px;
-  margin: 0 auto;
-`;
-
-const FoodListWrapper = styled.div`
-  display: flex;
-  padding: 10px;
-  gap: 10px;
-  flex-flow: row wrap;
-`;
-
 const Section = styled.section`
   max-width: 1020px;
   margin: 0 auto;
@@ -143,8 +131,6 @@ const MenuSection = styled.div`
 `;
 
 export {
-  FoodListContainer,
-  FoodListWrapper,
   Section,
   MainBody,
   FoodTruckName,
