+++ frontend/src/pages/FoodList/FoodTrucks/styles.js
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+const FoodListContainer = styled.div`
+  max-width: 1020px;
+  margin: 0 auto;
+`;
+
+const FoodListWrapper = styled.div`
+  display: flex;
+  padding: 10px;
+  gap: 10px;
+  flex-flow: row wrap;
+`;
+
+export { FoodListContainer, FoodListWrapper };
